Use a ClassLoader for locating resources
Before this change, the VirtualFileSystem was using a `Class` as a beacon
to find resources. I suppose that the intent was to use `Class#getResource`,
but in practice, it's always `clazz.getClassLoader().getResource()` which
is used, which makes it redundant to use a class. This commit refactors
the code to allow declaring which _classloader_ to use for loading. This
is important because in some cases, we don't necessarily have a `Class` to
pass, but we _do_ have a classloader. This should also allow more advanced
scenarios with filtering classloaders for example.

Since the `Class` was actually never directly used, but only its classloader,
this change should have 0 impact for backwards compatibility.